backend/pyproject.toml
[tool.poetry]
name = "academyvault"
version = "0.1.0"
description = ""
authors = ["maaz-links <mzit3116@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
psycopg2-binary = "^2.9.11"
djangorestframework = "^3.16.1"
mysqlclient = "^2.2.8"
django-extensions = "^4.1"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.9.0"
stripe = "^14.4.1"
python-dotenv = "^1.2.2"


[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
isort = "^8.0.1"
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
